Implemented AkkaSpout, tweets saved to topic

+++ src/main/scala/com/alawresz/twitterment/storm/bolts/TweetSinkConfig.scala
@@ -5,6 +5,6 @@ case class TweetSinkConfig(
   clientId: String,
   topic: String,
   batchSize: Int,
-  lingerMs: Short,
+  lingerMs: Long,
   acks: String
 )
